Speed up get_containers_testing!
The main fix is that it uses kubectl labels to only include pods
that are owned by the user in the original json. This implementation
makes only one kubectl call. However, I found that it's the many jq
calls on big json files that are the bottleneck for a call on all
pods, which can probably be fixed by parsing a single jq with awk.