Handle host keys & streamline run_pod background processes
Having a second potentially slow process to run after starting a
new pod made it clear that this approach is more maintainable.
Processes that should carry on after creating the pod are now funcions
in run_pod rather than external script(s). For pods with ssh access,
host keys are retrieved in a tmpfile so that they can be quickly read
by get_containers